requirements.txt
Django==2.2.5
django-cors-headers==3.1.0
django-filter==2.2.0
django-filters==0.2.1
djangorestframework==3.10.3
gunicorn==19.9.0
Pillow==6.1.0
pytz==2019.2
sqlparse==0.3.0
